# NeuronAgent/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: NeuronAgent API
  description: |
    REST API for NeuronAgent - AI agent runtime system providing agent state management,
    session handling, message processing, tool execution, and streaming capabilities.
  version: 1.0.0
  contact:
    name: NeuronDB Support
    email: support@neurondb.ai
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.neurondb.ai/v1
    description: Production server
...
paths:
...
  /agents:
    get:
      tags:
        - Agents
      summary: List agents
      description: List all agents, optionally filtered by search query
      operationId: listAgents
      parameters:
        - name: search
          in: query
          description: Search query to filter agents
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Agents
      summary: Create agent
      description: Create a new agent
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /tools:
    get:
      tags:
        - Tools
      summary: List tools
      description: List all available tools
      operationId: listTools
      responses:
        '200':
          description: List of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tool'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Tools
      summary: Create tool
      description: Create a new tool
      operationId: createTool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateToolRequest'
      responses:
        '201':
          description: Tool created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tools/{name}:
    get:
      tags:
        - Tools
      summary: Get tool
      description: Get tool by name
      operationId: getTool
      parameters:
        - name: name
          in: path
          required: true
          description: Tool name
          schema:
            type: string
      responses:
        '200':
          description: Tool details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Tools
      summary: Update tool
      description: Update an existing tool
      operationId: updateTool
      parameters:
        - name: name
          in: path
          required: true
          description: Tool name
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateToolRequest'
      responses:
        '200':
          description: Tool updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Tools
      summary: Delete tool
      description: Delete a tool by name
      operationId: deleteTool
      parameters:
        - name: name
          in: path
          required: true
          description: Tool name
          schema:
            type: string
      responses:
        '204':
          description: Tool deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key authentication using Bearer token
...
  schemas:
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Agent UUID
        name:
          type: string
          description: Agent name
        description:
          type: string
          nullable: true
          description: Agent description
        system_prompt:
          type: string
          description: System prompt for the agent
        model_name:
          type: string
          description: LLM model name
        memory_table:
          type: string
          nullable: true
          description: Memory table name
        enabled_tools:
          type: array
          items:
            type: string
          description: List of enabled tool names
        config:
          type: object
          additionalProperties: true
          description: Agent configuration
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - system_prompt
        - model_name
        - enabled_tools
        - config
        - created_at
        - updated_at
    CreateAgentRequest:
      type: object
      properties:
        name:
          type: string
          description: Agent name
        description:
          type: string
          nullable: true
          description: Agent description
        system_prompt:
          type: string
          description: System prompt for the agent
        model_name:
          type: string
          description: LLM model name
        memory_table:
          type: string
          nullable: true
          description: Memory table name
        enabled_tools:
          type: array
          items:
            type: string
          description: List of enabled tool names
        config:
          type: object
          additionalProperties: true
          description: Agent configuration (temperature, max_tokens, etc.)
      required:
        - name
        - system_prompt
        - model_name
        - enabled_tools
        - config
...
    Tool:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        schema:
          type: object
          additionalProperties: true
          description: JSON schema for tool parameters
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - name
        - description
        - schema
    CreateToolRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        schema:
          type: object
          additionalProperties: true
        implementation:
          type: string
          description: Tool implementation code or configuration
      required:
        - name
        - description
        - schema
    UpdateToolRequest:
      type: object
      properties:
        description:
          type: string
        schema:
          type: object
          additionalProperties: true
        implementation:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        code:
          type: integer
          description: HTTP status code
      required:
        - error
        - code
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Bad Request"
            message: "Invalid request parameters"
            code: 400
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not Found"
            message: "Resource not found"
            code: 404
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal Server Error"
            message: "An internal error occurred"
            code: 500
security:
  - BearerAuth: []